refactor: migrate context hooks over to hooks/useFetch
