+++ webui/react/src/App.tsx
@@ -5,8 +5,8 @@ import { setupAnalytics } from 'Analytics';
 import Link from 'components/Link';
 import Navigation from 'components/Navigation';
 import Router from 'components/Router';
-import { useFetchInfo } from 'contexts/Info';
 import StoreProvider, { useStore } from 'contexts/Store';
+import { useFetchInfo } from 'hooks/useFetch';
 import useKeyTracker from 'hooks/useKeyTracker';
 import usePolling from 'hooks/usePolling';
 import useResize from 'hooks/useResize';
