+++ webui/react/src/pages/TaskList.tsx
@@ -16,8 +16,8 @@ import TableBatch from 'components/TableBatch';
 import TaskActionDropdown from 'components/TaskActionDropdown';
 import TaskFilter from 'components/TaskFilter';
 import { useStore } from 'contexts/Store';
-import { useFetchUsers } from 'contexts/Users';
 import handleError, { ErrorLevel, ErrorType } from 'ErrorHandler';
+import { useFetchUsers } from 'hooks/useFetch';
 import usePolling from 'hooks/usePolling';
 import useStorage from 'hooks/useStorage';
 import { paths } from 'routes/utils';
